Fixed a bug that goobered the program when the a device's property included a float, now the devices will really display all their data.
Changed the default resolution option, so it will take full advantage of the running device's screen.

Attempted to get a service up and running, but unfortunately I based my service off of the Android 2.0 Library, and I know a lot of developers for opensource platforms like beagleboard are still working on 1.6, so I had to remove it for now just in case they one day decide to use this.